Backend_py/requirements.txt
fastapi
uvicorn
datetime
PyJWT
passlib
pydantic-settings
sqlalchemy
psycopg2
python-dotenv
pydantic
pydantic[email]
bcrypt